+++ src/core/grdc_encoding_options.h
@@ -10,19 +10,11 @@ typedef enum
     GRDC_ENCODING_MODE_RFX
 } GrdcEncodingMode;
 
-typedef enum
-{
-    GRDC_ENCODING_QUALITY_HIGH,
-    GRDC_ENCODING_QUALITY_MEDIUM,
-    GRDC_ENCODING_QUALITY_LOW
-} GrdcEncodingQuality;
-
 typedef struct
 {
     guint width;
     guint height;
     GrdcEncodingMode mode;
-    GrdcEncodingQuality quality;
     gboolean enable_frame_diff;
 } GrdcEncodingOptions;